# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_all",
        "summary": "List users",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/role/{role}": {
      "get": {
        "operationId": "UsersController_byRole",
        "summary": "List users by role",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_one",
        "summary": "Get user by id",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/register": {
      "post": {
        "operationId": "UsersController_register",
        "summary": "Public registration endpoint",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/portfolio": {
      "get": {
        "operationId": "UsersController_portfolio",
        "summary": "Get learner portfolio, holdings and trades",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/trades/user/{userId}": {
      "get": {
        "operationId": "TradesController_byUser",
        "summary": "List learner trades",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Trades"
        ]
      }
    },
...
    "/api/notifications/user/{userId}": {
      "get": {
        "operationId": "NotificationsController_byUser",
        "summary": "List notifications by user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "IndAI Review-4 Paper Trading API",
    "description": "NestJS modular backend with in-memory arrays, RBAC through x-role header, DTO validation, Swagger and minute-stock data endpoints.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "New",
            "minLength": 3,
            "description": "Required. Alphabets and spaces only."
          },
          "lastName": {
            "type": "string",
            "example": "Learner",
            "minLength": 3,
            "description": "Required. Alphabets and spaces only."
          },
          "email": {
            "type": "string",
            "example": "new@gmail.com",
            "description": "Required. Gmail or .in email only."
          },
          "password": {
            "type": "string",
            "example": "Test@1234",
            "minLength": 8,
            "description": "Required. Min 8 chars, uppercase, lowercase, number, and one of !@#$%^&*."
          },
          "confirmPassword": {
            "type": "string",
            "example": "Test@1234",
            "description": "Optional compatibility field used by frontend/Swagger. If provided, it must match password."
          },
          "role": {
            "type": "string",
            "enum": [
              "superuser",
              "admin",
              "instructor",
              "provider",
              "learner",
              "SUPER_USER",
              "ADMIN",
              "INSTRUCTOR",
              "COURSE_PROVIDER",
              "LEARNER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending",
              "suspended",
              "disabled"
            ],
            "example": "active"
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Optional but if provided must be 10 digits starting with 6/7/8/9."
          },
          "dateOfBirth": {
            "type": "string",
            "example": "2007-05-09",
            "description": "Required by registration frontend. Learner must be at least 10 years old."
          },
          "institution": {
            "type": "string",
            "example": "IIIT Sri City",
            "minLength": 3
          },
          "studentId": {
            "type": "string",
            "example": "CSE2026",
            "minLength": 3
          },
          "grade": {
            "type": "string",
            "example": "BTech 2nd Year",
            "minLength": 3
          },
          "major": {
            "type": "string",
            "example": "Computer Science",
            "minLength": 3
          },
          "organization": {
            "type": "string",
            "example": "IndAI Academy",
            "minLength": 3
          },
          "expertise": {
            "type": "string",
            "example": "Technical Analysis",
            "minLength": 3
          },
          "yearsExp": {
            "type": "string",
            "example": "3-5"
          },
          "website": {
            "type": "string",
            "example": "https://example.com"
          },
          "contentType": {
            "type": "string",
            "example": "Video Courses"
          },
          "authCode": {
            "type": "string",
            "example": "ADMIN-2026"
          },
          "accessLevel": {
            "type": "string",
            "example": "Full"
          },
          "tradingExperience": {
            "type": "string",
            "enum": [
              "Beginner",
              "Intermediate",
              "Advanced",
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "0-1",
              "1-3",
              "3-5",
              "5+"
            ]
          },
          "riskTolerance": {
            "type": "string",
            "enum": [
              "Low",
              "Medium",
              "High",
              "Conservative",
              "Moderate",
              "Aggressive",
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "startingBalance": {
            "type": "number",
            "example": 100000
          },
          "tradingLimit": {
            "type": "number",
            "example": 150000
          },
          "instructorId": {
            "type": "string",
            "example": "u3"
          },
          "skillPoints": {
            "type": "number",
            "example": 50
          },
          "portfolioValue": {
            "type": "number",
            "example": 100000
          },
          "virtualBalance": {
            "type": "number",
            "example": 100000
          },
          "experience": {
            "type": "string",
            "example": "Beginner"
          },
          "goals": {
            "type": "string",
            "example": "Learn paper trading safely"
          },
          "studentIds": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "yearsOfExperience": {
            "type": "number",
            "example": 5,
            "description": "Compatibility alias for yearsExp used by some Swagger/frontend payloads."
          },
          "experienceLevel": {
            "type": "string",
            "enum": [
              "Beginner",
              "Intermediate",
              "Advanced",
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "0-1",
              "1-3",
              "3-5",
              "5+"
            ],
            "description": "Compatibility alias for experience/tradingExperience."
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending",
              "suspended",
              "disabled"
            ]
          },
          "phone": {
            "type": "string"
          },
          "tradingLimit": {
            "type": "number"
          },
          "skillPoints": {
            "type": "number"
          },
          "instructorId": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "major": {
            "type": "string"
          }
        }
      },
...
    }
  }
}